// openapi-orval/front/src/api/generated.ts
export type PostApiUsersBody = {
  /** @minLength 1 */
  name: string;
  email: string;
};